backend/requirements.txt
crewai
crewai-tools
flask
flask-cors
python-dotenv
pydantic
anthropic
litellm
numpy
pandas